Handle Category & Tag deleted which associated to trashed posts add hasCategory in Post model add hasTrashedPosts in Tag and Category model

+++ app/Http/Controllers/CategoriesController.php
@@ -111,6 +111,18 @@ class CategoriesController extends Controller
     public function destroy(Category $category)
     {
         //
+        if($category->posts->count() > 0)
+        {
+            session()->flash('error', 'Category cannot be deleted; because it has some posts.');
+            return redirect()->back();
+        }
+
+        if($category->hasTrashedPosts())
+        {
+            session()->flash('error', 'Category cannot be deleted; because it has some trashed posts.');
+            return redirect()->back();
+        }
+
         $category->delete();
 
         session()->flash("success", "Category deleted successfully");
