+++ app/Http/Controllers/TagsController.php
@@ -107,6 +107,18 @@ class TagsController extends Controller
     public function destroy(Tag $tag)
     {
         //
+        if($tag->posts()->count() > 0)
+        {
+            session()->flash('error', 'Tag cannot be deleted; because it is associated to posts.');
+            return redirect()->back();
+        }
+        // dd($tag->hasTrashedPosts());
+        if($tag->hasTrashedPosts())
+        {
+            session()->flash('error', 'Tag cannot be deleted; because it is associated to trashed posts.');
+            return redirect()->back();
+        }
+
         $tag->delete();
 
         session()->flash("success", "Tag deleted successfully");
